Fix adding buildkeychain flag multiple times to OTHER_CODE_SIGN_FLAGS
Description
===========

Small fix for an issue which made no real harm at runtime at the moment
but needs a fix in any case. The `Provider` for build arguments in
`XcodeArchive` is mutating the `Buildsettings` object of the task
while constructing the argument `List`. The provider itself
gets called multiple times and manipulates the BuildSettings again
and again and again. I decided to go with the `clone` approuch here
instead of memorizing the whole closure. First I'm not always sure if
the first invocation is already late enough so that all fields are
set correctly.

While testing this I saw that my initial `clone` implementation in
`BuildSettings` had a major bug. I also need to clone the underlying
`HashMap` which holds all passed values.
This fix allows to call `buildArguments.get()` multiple times without
multiplying the keychain flags

Changes
=======

* ![FIX] adding buildkeychain flag multiple times to OTHER_CODE_SIGN_FLAGS