Improve task output caching
Description
===========

The task output caching was practically not working due to the
task setup. This patch makes sure that the `importCodeSigningIdentities`
can be cached and the xcode build tasks run only when the sources change.

I can not setup a proper test for this yet as we would need credentials
and certificates set up as well. I will create those in a later patch.

Changes
=======

* ![IMPROVE] task output caching