Implement continueOnFailure passthrough/property for GradleBuild task
Description
===========

This patch adds a new input property `continueOnFailure` which
will set a the matching property on the `GradleBuild` launcher.
If this property is not set, the task will check the start parameter
of the running gradle instance if it has the flag set which means it
will by default pass the flag through analog to the loglevel.

If the `--continue` flag is already set in the tasks buildarguments
it will be used instead.

Changes
=======

* ![IMPROVE] `GraddleBuild` continueOnFailure flag/passthrough