Improve net.wooga.build-unity-ios xcode task setup
Description
===========

This patch changes drastically how the xcode related tasks are created,
named and managed. The old implementation scanned the whole
project directory tree for `.xcodeproj` directories. The idea was to
be able to manage multiple projects with one plugin. I changed
how this works to simplify the setup. In our opinioted view only
a single xcode project is located in the project directory.
We set a default value to `Unity-iPhone` as the base name for
this project. This and the xcode project directory property
can be set in the extension. We now only setup tasks for this
one project. This makes the base setup way simpler.

I'm still not happy how to handle xcodeproj and xcworkspace directories.
I deciced to go the `xcodeProjectPath` + `xcodeProjectBaseName` + `extension` route.
The downside is that one cannot simply set the path to an xcode project. In the
old API without providers this would have been possible. But the convention
mapping only really works from bottom to the top.

This should not be so dramatic at the moment.

Changes
=======

* ![IMPROVE] `net.wooga.build-unity-ios` xcode task setup
* ![IMPROVE] PodInstall task properties and test setup