Add export lifecycle task and skipExport logic (#145)
Description
===========

The `net.wooga.build-unity` plugin defined dependencies
to the basic lifecycle tasks `assemble, build, check, publish)
when the `defaultAppConfigName` property was provided.
The only missing key is the `export` method.

This patch adds the missing link and allows to call `export`
on a project with `defaultAppConfigName` set to easily run
and execute the correct `export` variant.

I also added a way of skipping the export action. This might become
usefull/needed for teams who need to enforce that the export will be
skipped. At the moment this has to be done since most Unity projects
still alter the source files when building.

Changes
=======

* ![ADD] lifecycle task `export`
* ![ADD] `skipExport` logic for export task actions