Improve exportOptions.plist handling (#146)
Description
===========

This patch adds a new way to configure the export options plist
file needed during the archive command run.
A lot of information we need for a complete build is tracked
multiple times. This patch tries to make it easier to define these
config values only once. On the other hand it gives a new way of
configuring the content of the exportOptions.plist file in the
build.gradle file.

The first change is the introduction of a new provider in
the ios plugin extension:

`Provider<ExportOptions> getExportOptions()`

With this provider one can read and use the values in the
`exportOptions.plist` file at runtime without additional parsing.

With this provider created I set some conventions to point to the matching
value in this new provider

```groovy
  extension.teamId.convention(extension.exportOptions.map({ it.teamID }))
  extension.signingIdentities.convention(extension.exportOptions.map({it.signingCertificate ? [it.signingCertificate] : []}).orElse(project.provider({ new ArrayList<String>() })))
  extension.adhoc.convention(extension.exportOptions.map({ it.method == 'ad-hoc' }).orElse(false))
  extension.appIdentifier.convention(extension.exportOptions.map({ it.distributionBundleIdentifier }))
```

Means that `teamId`, `adhoc`, `appIdentifier` and the initial value for `signingIdenties`
are all provided through the contents of the `exportOption.plist` file.
These values can of course be overriden in the extension when needed.

On top of that, one can also issue additional overrides.
This works only with configuration actions/closures due to the nature
of the lazy evalution of the provider itself.

```groovy
iosBuild {
  exportOptions {
    teamID = "my team ID"
    method = 'ad-hoc'
  }
}
```

In this example the configuration action/closure would be saved and
applied right before the value gets evaluated. It will also
create a copy of the provided `exportOptions.plist` file and
copy it along with the adjusted values to a new location within the
`build` directory. For this I added yet another internal provider
to the extension:

```groovy
Provider<RegularFile> getFinalExportOptionsPlist() {
  finalExportOptionsPlist.present ? finalExportOptionsPlist : baseExportOptionsPlist
}
```

It will either return the original plist file or a copy with
adjusted values when configurations where registered in the `build.gradle` file.

I added some more documentation strings to the `ExportOptions` class and
improved the `ExportOptions.DistributionManifest` constructor methods.

Changes
=======

* ![ADD] `exportOptions` provider in plugin extension
* ![IMPROVE] conventions for values declared through `exportOption.plist`
* ![ADD] logic to lazy configure export options and merge values into new config