Remove coveralls reporting (#156)
Description
===========

Because of security issues reported for the
coveralls gradle plugin I decided to remove the feature
during build. It is not enough to just stop sending
reports via jenkins. We also need to update to
`net.wooga.plugins` > `4.0.0` because this version removes
the dependency.

I remove the coveralls token in the Jenkinsfile so
our base pipeline won't try to call the coveralls task
on the gradle project.

Changes
=======

* ![UPDATE] `net.wooga.plugins` to version `4.0.0` to remove coveralls dependency
* ![REMOVE] coveralls token in Jenkinsfile