Improve build keychain default settings
Description
===========

A simple patch to set some basic conventions for the default
settings for the build keychain. Without these the keychain
would be created with the default settings (lock after sleep and
lock after 5 minutes) This will ensure that longer builds still can
access the keychain after a compile.

I also set a domain value for the keychain operation to not
write into the system values etc. This is just a precauthion since the
user in question can only set user keychains anyway. But the commandline
arguments can be smaller and it is in general a better way.

Changes
=======

* ![IMPROVE] default build keychain settings
* ![IMPROVE] set domain value for keychain searchlist operations