Improve multiple provisioning profiles support
Description
===========

This patch adds a new fastlane task type `SighRenewBatch` which
allows to set a map of `bundleId` -> `profileName`.
This allows to set the map pulled from the `exportOption.plist` into
the `importProvisioningProfiles` task.

I'm not 100% happy with the implementation but one could revisit this
with another patch to rework the inner workings.

At the moment the new task type simply extends `SighRenew` and
adds a `profiles` property which gets looped over. The task action
of `SighRenew` is called multiple times. This has the desired effect
but can cause issues with input caching etc. I'm fine with this from
a first version prespective

Changes
=======

* ![IMPROVE] multiple provisioning profiles support