Move PropertyUtils to private static class (#165)
Description
===========

The static class `wooga.gradle.build.unity.ios.internal.utils.PropertyUtils` was
crossreferenced from other namespaces. We could move the functionality
to gradle-commons but I'm not 100% sure if the logic still needed.
I opted to duplicate the class in both plugin namespaces that use it because
we will split this plugin into multiple repos. The logic is not
so complicated.

Changes
=======

* ![IMPROVE] `PropertyUtils` as private static class