Add empty plugin class for net.wooga.macos-security plugin (#167)
Description
===========

For the split of the internal plugin now named `net.wooga.macos-security` we need
to declare a plugin class. The plugin itself brings in no auto added tasks or
conventions at the moment. This is the reason why the plugin is empty. I have
some ideas about use case conventions but I want to keep the plugin as generic
as possible for the time being.

Changes
=======

* ![ADD] ![GRADLE] plugin class for `net.wooga.macos-security`