Improve fastlane tasks with ExecSpec
Description
===========

This patch adds the new `ExecSpec` and `ProcessExecutor` utils
to execute `fastlane` on the host systems. The new tools allow
to set the executable path and executable name to be set individually.
It also allows us to easily override the path to fastlane in the tests.

The `ProcessExecutor` helper is a new utility class to execute
executables with a more convinient API for our internal usecases with
logging, arguments etc.

Changes
=======

* ![IMPROVE] `fastlane` tasks with `ExecSpec` and `ProcessExecutor`