Improve net.wooga.build-unity-ios with configuration avoidance API
Description
===========

The plan is to split the `net.wooga.build-unity-ios` plugin out of this repo.
I looked over the Plugin class and saw that we still used the static task creation
setup. I switched all over to the `tasks.register` methods and updated
the calls to work with this new setup. The speed gain should not be super high
because we need all tasks in the run anyway but it is the cleaner setup overall.

Changes
=======

* ![IMPROVE] `net.wooga.build-unity-ios` plugin with configuration avoidance API