Add UBS 1.6 compatibility setup
Description
===========

With UBS 1.6 and especially in combination with Unity > 2021 we have
to introduce some version specific changes.
The cli parameter `--version` is no longer valid and when provided in the
cli call to Unity will result in Unity printing its version.
So I renamed the cli flag to `--build-version` when UBS compatibility is set to
`v160`.

> Note:
> `--version` was already not working in UBS because of an issue with the used
> CLI parser. I kept the old behavior just to be compatible to the old version.

I also renamed `--version-code` to `--build-version-code`.

While touching things based on the UBS compatibility value I also changed
the default `exportMethod` value and `appConfigsDirectory`.

Changes
=======

* ![ADD] UBS 1.6 compatibility setup